Initial work in producing random data into Kafka
This is the first script i have written in GO. Currently it expects the kafka server to be up and will push the data accordingly.

There are numerous flaws in the program including graceful shutdown and too much CPU usage.  I also should reduce the data being produced and make it configurable.